team_recommendation_api/src/requirements.txt
fastapi
uvicorn
google-generativeai
psycopg2
python-dotenv
pydantic
